GenerativeAIExamples/Projects/Gemini_using_MySQL_DB_LLM/requirements.txt
mysql-connector-python
pymysql
sentence-transformers
chromadb
mysql
flask
google-generativeai
plotly
plotly.express
pandas
mysql-connector
python-dotenv
seaborn
matplotlib
langchain
pypdf
IPython
llama-index-embeddings-gemini
streamlit